new props will rerender the tree
